Added samples and tested it

+++ Solver/sudoku_solver.c
@@ -324,7 +324,7 @@ int str_is_equal(const char *a, const char *b)
     unsigned long len = strlen(a);
     if(len != strlen(b))
         return false;
-    for(int i = 0; i <len; i++)
+    for(unsigned long i = 0; i <len; i++)
     {
         if(a[i] != b[i])
             return false;
